+++ entity/src/main/java/org/grnet/cat/entities/Assessment.java
@@ -57,5 +57,6 @@ public class Assessment {
                 cascade = CascadeType.ALL,
                 orphanRemoval = true)
     private List<Comment> comments = new ArrayList<>();
+    private Boolean shared;
 
 }